dynamic-arrays.cpp
This code allows a user to enter the size of the array as well as the value. THe code allocates memory for an array with a pointer and prints out the values. At the end, the array is deallocated and set to null. This is done because the memory allocation was only temporary. If the values were not removed, it could cause a lot of issues when re-runnning the program.